+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   and will cast it into an array.
 - Add total runtime metric method and refine time measurement. You can easily get the `Metric` object of the total
   runtime via `C::Performance()->getTotalRuntimeMetric()` and easily adjust the start and end time of `Metric` objects.
+- When an error occurs while outputting the error exception will be thrown instead of a generic one. This gives
+  you the correct stack trace for debugging, if exception throwing is enabled.
 
 ### 🐞 Fixed
 - The database migrator command `bob db:sync` now displays the stats correctly.