+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### 🐞 Fixed
 - The database migrator command `bob db:sync` now displays the stats correctly.
+- Command's `$this->io->choice(...)` worke fine again.
 
 ### 🔥 Removed
 - Dropped support for FTP filesystems (SFTP is preferred if needed at all).